Moved some components to typescript
