+++ src/components/search/results/TimelineView.tsx
@@ -2,7 +2,10 @@ import React, { useEffect } from "react";
 import { useQuery } from "@apollo/client";
 import moment from "moment";
 
-import { Masonry } from "@mui/lab";
+import { Masonry, Timeline, TimelineItem, TimelineDot } from "@mui/lab";
+import { TimelineSeparator, TimelineConnector } from "@mui/lab";
+import { TimelineContent, TimelineOppositeContent } from "@mui/lab";
+
 import { Box, Paper, Typography } from "@mui/material";
 import { AccordionDetails, AccordionSummary } from "@mui/material";
 
@@ -14,14 +17,6 @@ import { GET_INTERACTIONS } from "../../../queries";
 import resultsState from "../../../datastore/resultsStore";
 import StyledAccordion from "../../generic/StyledAccordion";
 import InteractionsCard from "./InteractionsCard";
-
-import Timeline from '@mui/lab/Timeline';
-import TimelineItem from '@mui/lab/TimelineItem';
-import TimelineSeparator from '@mui/lab/TimelineSeparator';
-import TimelineConnector from '@mui/lab/TimelineConnector';
-import TimelineContent from '@mui/lab/TimelineContent';
-import TimelineOppositeContent from '@mui/lab/TimelineOppositeContent';
-import TimelineDot from '@mui/lab/TimelineDot';
 
 type InteractionsListProps = {
   searchTerm: String;
@@ -57,108 +52,107 @@ export default function TimelineView(props: InteractionsListProps) {
           gap: "0.5em",
           padding: "0.5em",
           width: "100%",
-      }}>
-
-      <Timeline position="alternate">
-      {interactions.map(
-        (
-          {
-            pk,
-            sk,
-            case_ref,
-            exhibit_ref,
-            device_uid,
-            file_hash,
-            organisation,
-            datetime_added,
-            datetime,
-            local_partner,
-            interaction,
-            direction,
-            duration,
-            status,
-          },
-          index
-        ) => (
-        <TimelineItem>
-          <TimelineOppositeContent
-            sx={{ m: 'auto 0' }}
-            align="right"
-            variant="body2"
-            color="white"
-          >
-            {moment(datetime).format("MMM Do YYYY, h:mma")}
-          </TimelineOppositeContent>
-          <TimelineSeparator>
-            <TimelineConnector />
-            <TimelineDot>
-              <Box
-                sx={{
-                  float: "left",
-                }}
-              >
-                {status === "missed" && (
-                  <PhoneDisabledIcon
-                    fontSize="medium"
-                    sx={{ color: "red", marginBottom: -0.6 }}
-                  />
-                )}
-                {status === "answered" && (
-                  <PhoneIcon
-                    fontSize="medium"
-                    sx={{ color: "green", marginBottom: -0.6 }}
-                  />
-                )}
-              </Box>
-            </TimelineDot>
-            <TimelineConnector />
-          </TimelineSeparator>
-          <TimelineContent>
-            <Masonry columns={1} spacing={2}>
-              <Paper key={index}>
-                <StyledAccordion>
-                  <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-                     <Typography>
-                      <Box
-                        sx={{
-                          float: "right",
-                        }}
-                      >
-                        &nbsp;&nbsp;{index + 1} - {status} ({duration})
-                        <br />
-                        {moment(datetime).format("MMM Do YYYY, h:mma")}
-                      </Box>
-                    </Typography>
-                  </AccordionSummary>
-                  <AccordionDetails>
-                    <InteractionsCard
-                      key={index}
-                      pk={pk}
-                      sk={sk}
-                      case_ref={case_ref}
-                      exhibit_ref={exhibit_ref}
-                      device_uid={device_uid}
-                      file_hash={file_hash}
-                      organisation={organisation}
-                      datetime_added={datetime_added}
-                      datetime={datetime}
-                      local_partner={local_partner}
-                      interaction={interaction}
-                      direction={direction}
-                      duration={duration}
-                      status={status}
-                    />
-                  </AccordionDetails>
-                </StyledAccordion>
-              </Paper>
-            </Masonry>
-          </TimelineContent>
-        </TimelineItem>
-        )
-      )}
-      </Timeline>
-    </Box>
-
+        }}
+      >
+        <Timeline position="alternate">
+          {interactions.map(
+            (
+              {
+                pk,
+                sk,
+                case_ref,
+                exhibit_ref,
+                device_uid,
+                file_hash,
+                organisation,
+                datetime_added,
+                datetime,
+                local_partner,
+                interaction,
+                direction,
+                duration,
+                status,
+              },
+              index
+            ) => (
+              <TimelineItem>
+                <TimelineOppositeContent
+                  sx={{ m: "auto 0" }}
+                  align="right"
+                  variant="body2"
+                  color="white"
+                >
+                  {moment(datetime).format("MMM Do YYYY, h:mma")}
+                </TimelineOppositeContent>
+                <TimelineSeparator>
+                  <TimelineConnector />
+                  <TimelineDot>
+                    <Box
+                      sx={{
+                        float: "left",
+                      }}
+                    >
+                      {status === "missed" && (
+                        <PhoneDisabledIcon
+                          fontSize="medium"
+                          sx={{ color: "red", marginBottom: -0.6 }}
+                        />
+                      )}
+                      {status === "answered" && (
+                        <PhoneIcon
+                          fontSize="medium"
+                          sx={{ color: "green", marginBottom: -0.6 }}
+                        />
+                      )}
+                    </Box>
+                  </TimelineDot>
+                  <TimelineConnector />
+                </TimelineSeparator>
+                <TimelineContent>
+                  <Masonry columns={1} spacing={2}>
+                    <Paper key={index}>
+                      <StyledAccordion>
+                        <AccordionSummary expandIcon={<ExpandMoreIcon />}>
+                          <Typography>
+                            <Box
+                              sx={{
+                                float: "right",
+                              }}
+                            >
+                              &nbsp;&nbsp;{index + 1} - {status} ({duration})
+                              <br />
+                              {moment(datetime).format("MMM Do YYYY, h:mma")}
+                            </Box>
+                          </Typography>
+                        </AccordionSummary>
+                        <AccordionDetails>
+                          <InteractionsCard
+                            key={index}
+                            pk={pk}
+                            sk={sk}
+                            case_ref={case_ref}
+                            exhibit_ref={exhibit_ref}
+                            device_uid={device_uid}
+                            file_hash={file_hash}
+                            organisation={organisation}
+                            datetime_added={datetime_added}
+                            datetime={datetime}
+                            local_partner={local_partner}
+                            interaction={interaction}
+                            direction={direction}
+                            duration={duration}
+                            status={status}
+                          />
+                        </AccordionDetails>
+                      </StyledAccordion>
+                    </Paper>
+                  </Masonry>
+                </TimelineContent>
+              </TimelineItem>
+            )
+          )}
+        </Timeline>
+      </Box>
     );
   }
 }
